moved expandable operating status to shared components
