+++ src/templates/components/expandableOperatingStatus/index.tsx
